requirements.txt
requests~=2.28.1
ddddocr<=1.1.0
retry~=0.9.2
beautifulsoup4~=4.11.1
pillow~=9.3.0
click~=8.1.3